update all schemas in layer
